Fixed application-test.properties Added registration controller with service request and response Tested registration service Added ErrorController Tested Error controller Added migration for users table for dev

+++ src/main/resources/db/dev/migration/V1_1__init_urls_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE urls(
                      id BIGINT PRIMARY KEY,
                      short_url VARCHAR(12) NOT NULL ,
                      long_url VARCHAR(256) NOT NULL ,
-                     user_id INTEGER REFERENCES users(id),
+                     user_id BIGINT REFERENCES users(id),
                      visited BIGINT DEFAULT 0,
                      created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                      expired_at TIMESTAMP DEFAULT (SELECT DATEADD('DAY', 14, CURRENT_TIMESTAMP))
