+++ src/test/resources/db/test/migration/V1__init_db.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS urls(
                                    id BIGSERIAL PRIMARY KEY,
                                    short_url VARCHAR(12) NOT NULL ,
                                    long_url VARCHAR(256) NOT NULL ,
-                                   user_id INTEGER REFERENCES users(id),
+                                   user_id BIGINT REFERENCES users(id),
                                    visited BIGINT DEFAULT 0,
                                    created_at TIMESTAMP DEFAULT current_timestamp,
                                    expired_at TIMESTAMP DEFAULT current_timestamp + INTERVAL '14 days'
